backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
groq==0.9.0
httpx==0.27.2
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.7.4
nltk==3.8.1
beautifulsoup4==4.12.3
aiofiles==23.2.1
